+++ app/Http/Controllers/PenjualanController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class PenjualanController extends Controller
+{
+    public function index()
+    {
+        return view('datapenjualan');
+    }
+}
